use auth cookies to avoid switch max sessions failure
Add capability to reuse same session for each request. Refer NX-OS_Programmability_Guide
NX-API provides a session-based cookie, nxapi_auth when users first
successfully authenticate. With the session cookie, the username and
password are included in all subsequent NX-API requests that are sent to
the device. The username and password are used with the session cookie
to bypass performing the full authentication process again. If the
session cookie is not included with subsequent requests, another session
cookie is required and is provided by the authentication process.
Avoiding unnecessary use of the authentication process helps to reduce
the workload on the device.

Signed-off-by: Leslie Qi Wang <qiwa@pensando.io>